==> notebooks/15_make_sensitivity_specificity_table.Rmd <==
---
title: "`r gsub('.Rmd', ' ', gsub('_', ' ', knitr::current_input()))`"
author: "Holly Beale"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
  # pdf_document: default
  # fig_caption: yes
---

```{r setup, include = FALSE}


library(readr)
library(tidyr)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(knitr)
library(forcats)

```

```{r}

neighbor_sens_spec_raw <- read_tsv("../data/neighbor_accuracy_stats_combo_for_pub.tsv") %>% 
  dplyr::rename(bin_median = median_of_depth_bin) %>%
  mutate(bin_median = ifelse(bin_median == 0, 1, bin_median))

outlier_sens_spec_raw <- read_tsv("../data/outlier_median_accuracy_stats_combo_for_pub.tsv") 

predicted_required_total_reads <- read_tsv("../data/predicted_required_total_reads_within_limits.tsv") %>%
  select(-percentile_to_query)

sens_spec <- left_join(outlier_sens_spec_raw,
                       neighbor_sens_spec_raw,
                       by = c(median_of_depth_bin = "bin_median")) %>%
  left_join(predicted_required_total_reads, c("median_of_depth_bin" = "proposed_threshold"))
  
colnames(sens_spec) = c("MEND reads (million)",
                        "Outlier sensitivity",
                        "Outlier specificity",
                        "Correlation sensitivity",
                        "Correlation specificity",
                        "Estimated total reads")

sens_spec %>%
  mutate_at(vars(contains('ity')), funs(sprintf("%0.4f",.))) %>%  
  # mutate(`Estimated total reads`= prettyNum(`Estimated total reads`, big.mark = ","))  %>%  
  mutate(`Estimated total reads` = round(`Estimated total reads` / 1E6, 1))  %>%  
  rename(`Estimated total reads (millions)` = `Estimated total reads`)  %>% 
  write_tsv("../figures_and_tables/table_2_sens_spec_and_thresholds.tsv")

```
