==> notebooks/20_make_supplemental_subsample_table.Rmd <==
---
title: "`r gsub('.Rmd', ' ', gsub('_', ' ', knitr::current_input()))`"
author: "Holly Beale & Jackie Roger"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document: default
  # pdf_document: default
  # fig_caption: yes
---

```{r setup, include = FALSE}

library(readr)
library(tidyr)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(knitr)
library(forcats)

```

## Load data

```{r}

read_counts_raw <- read_tsv("../data/read_counts.txt", 
                            col_names = c("sample_id", "measurement", "read_count"))


evenly_selected_subsamples <- read_tsv(paste0("../data/evenly_selected_subsamples_with_binwidth_4.tsv"))


parent_publication_ids <- read_tsv("../data/parent_sample_data_source.tsv") %>%
  select(Eval_ID, publication_id)

```

## Groom read counts

```{r}

read_counts_anno <- read_counts_raw %>% 
  filter(sample_id %in% evenly_selected_subsamples$sample_id) %>%
  mutate(target_MEND_count = as.numeric(gsub("^.*_est([0-9]*)M_.*$", "\\1", sample_id)),
         parent_id = gsub("_est.*$", "", sample_id)) %>%
  spread(measurement, read_count) %>%
  mutate(Mapped_read_count = Multimapped_read_count + Uniquely_mapped_read_count,
         MND = Mapped_non_duplicate_read_count,
         pct_MEND = MEND / total_sequences,
         not_mapped = total_sequences - Mapped_read_count,
         duplicate_reads = Mapped_read_count - MND,
         non_exonic_reads = MND - MEND,
         pct_MMoM = Multimapped_read_count/Mapped_read_count,
         seed = substr(sample_id, nchar(sample_id)-3,
                                        nchar(sample_id))
  )

```

## Make table with subsample fractions

```{r}

subsample_fractions <- read_counts_anno %>% 
  left_join(parent_publication_ids, by=c("parent_id"="Eval_ID")) %>%
  rename(Sample = publication_id) %>%
  select(-parent_id)

# subsample_fractions$MEND <- round(subsample_fractions$MEND, digits = 0)
# subsample_fractions$pct_MEND <- round(subsample_fractions$pct_MEND, digits = 1)

subsample_fractions2 <- subsample_fractions %>%
select(Sample, `Target MEND Count (M)` = target_MEND_count, Seed = seed, `Total sequences` = total_sequences, `Not mapped` = not_mapped, `Duplicates` = duplicate_reads, `Non-exonic` = non_exonic_reads, MEND,
         `Percent MEND` = pct_MEND, `Percent MMoM` = pct_MMoM) %>%
  mutate(`Non-exonic` = format(round(`Non-exonic`), big.mark = ","),
         `Percent MEND` = round(100 * `Percent MEND`,1),
         `Percent MMoM` = round(100 * `Percent MMoM`,1),
         `Total sequences` = format(`Total sequences`, big.mark = ","),
         `Not mapped` = format(`Not mapped`, big.mark = ","),
         `Duplicates` = format(round(`Duplicates`), big.mark = ","),
         `MEND` = format(round(`MEND`), big.mark = ",")
         ) %>%
  arrange(Sample, `Target MEND Count (M)`)
 
# [,c(2, 1, 7, 3, 6, 5, 4, 8)]

subsample_table <- subsample_fractions2 %>%
  write_tsv("../figures_and_tables/table_S2_subsamples.tsv")

```
